reorganize panel components by sidebar category (option A)
