+++ web/src/panels/control/PieceStatePanel.jsx
