+++ web/src/panels/control/RotationControlPanel.jsx
