+++ web/src/panels/perception/BuildingEditorPanel.jsx
