+++ web/src/panels/perception/VisionPanel.jsx
