+++ web/src/tabs/CubeTab.jsx
@@ -1,7 +1,7 @@
-/** Legacy compatibility layer: use panels/cube/* directly for new code. */
-import { CubeViewerPanel } from '../panels/cube/CubeViewerPanel';
-import { PieceStatePanel } from '../panels/cube/PieceStatePanel';
-import { RotationControlPanel } from '../panels/cube/RotationControlPanel';
+/** Legacy compatibility layer: use panels/control/* directly for new code. */
+import { CubeViewerPanel } from '../panels/control/CubeViewerPanel';
+import { PieceStatePanel } from '../panels/control/PieceStatePanel';
+import { RotationControlPanel } from '../panels/control/RotationControlPanel';
 
 export default function CubeTab() {
   return <div><CubeViewerPanel /><RotationControlPanel /><PieceStatePanel /></div>;
