+++ web/src/tabs/KnowledgeTab.jsx
@@ -1,6 +1,6 @@
-/** Legacy compatibility layer: use panels/knowledge/* directly for new code. */
-import { DocumentBrowserPanel } from '../panels/knowledge/DocumentBrowserPanel';
-import { BuildingEditorPanel } from '../panels/knowledge/BuildingEditorPanel';
+/** Legacy compatibility layer: use panels/perception/* directly for new code. */
+import { DocumentBrowserPanel } from '../panels/perception/DocumentBrowserPanel';
+import { BuildingEditorPanel } from '../panels/perception/BuildingEditorPanel';
 
 export default function KnowledgeTab() {
   return <div><DocumentBrowserPanel /><BuildingEditorPanel /></div>;
